move kusto queries to separate file
